Cast Transpose to int when adding to note.

+++ src/OPL3GM_proc.cpp
@@ -124,7 +124,7 @@ VstInt32 OPL3GM::processEvents (VstEvents* ev)
 					if (channel != 9)
 					{
 						float note = byte2;
-						note = note + Transpose;
+						note = note + (int)Transpose;
 						if (note > 127)
 						{
 							note = 127;
